+++ caffe2/operators/relu_op_cudnn.cc
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "caffe2/core/common_cudnn.h"
+#include "caffe2/core/cudnn_wrappers.h"
 #include "caffe2/core/context_gpu.h"
 #include "caffe2/core/operator.h"
 #include "caffe2/core/types.h"
