+++ caffe2/operators/spatial_batch_norm_op_cudnn.cc
@@ -16,7 +16,7 @@
 
 #include <cfloat>
 
-#include "caffe2/core/common_cudnn.h"
+#include "caffe2/core/cudnn_wrappers.h"
 #include "caffe2/core/context_gpu.h"
 #include "caffe2/operators/spatial_batch_norm_op.h"
 #include "caffe2/utils/math.h"
